throttle moved to plugins
