+++ lib/rocketjob.rb
@@ -57,7 +57,6 @@ module RocketJob
       autoload :Persistence,         "rocket_job/plugins/job/persistence"
       autoload :Throttle,            "rocket_job/plugins/job/throttle"
       autoload :ThrottleRunningJobs, "rocket_job/plugins/job/throttle_running_jobs"
-      autoload :ThrottleDependentJobs, "rocket_job/plugins/job/throttle_dependent_jobs"
       autoload :Transaction,         "rocket_job/plugins/job/transaction"
       autoload :Worker,              "rocket_job/plugins/job/worker"
     end
@@ -69,6 +68,7 @@ module RocketJob
     autoload :Singleton,             "rocket_job/plugins/singleton"
     autoload :StateMachine,          "rocket_job/plugins/state_machine"
     autoload :Transaction,           "rocket_job/plugins/transaction"
+    autoload :ThrottleDependentJobs, "rocket_job/plugins/throttle_dependent_jobs"
   end
 
   module Jobs
